fixed many imports, depictor refactoring
