+++ peepingtom/datablocks/datablock.py
@@ -3,11 +3,10 @@ class DataBlock:
     Base class for all simple and complex datablocks.
     Provides common methods and easy type inference
     """
-    _depictor_class = lambda: None
+    _depictor_class = None
 
     def __init__(self, name=None):
         self.name = name
- #       self.depictor = self._depictor_class(self)
 
     def __newlike__(self, *args, **kwargs):
         # this makes sure that operators get the right output in case
@@ -16,6 +15,12 @@ class DataBlock:
             return NotImplemented
         cls = type(self)
         return cls(*args, **kwargs)
+
+    def _init_depictor(self, depictor):
+        if depictor is not None:
+            return depictor
+        else:
+            return self._depictor_class(self)
 
     def __shape_repr__(self):
         return ''
