--- peepingtom/depictors/__init__.py
@@ -1,5 +1,0 @@
-from .base import Depictor
-from .images import ImageDepictor
-from .lines import LineDepictor
-from .particles import ParticleDepictor
-from .points import PointDepictor
