+++ peepingtom/depictors/blocks/__init__.py
@@ -1,0 +1,5 @@
+from .depictor import Depictor
+from .imagedepictor import ImageDepictor
+from .linedepictor import LineDepictor
+from .particledepictor import ParticleDepictor
+from .pointdepictor import PointDepictor
