+++ peepingtom/depictors/blocks/imagedepictor.py
