+++ peepingtom/depictors/blocks/linedepictor.py
