+++ peepingtom/depictors/blocks/particledepictor.py
