+++ peepingtom/depictors/blocks/pointdepictor.py
