--- peepingtom/depictors/dipoles.py
@@ -1,4 +1,0 @@
-from .base import Depictor
-
-class DipoleDepictor(Depictor):
-    pass
