--- peepingtom/depictors/peeper/__init__.py
@@ -1,1 +1,0 @@
-from .base import Peeper, peep
