+++ peepingtom/io_/read/mrc/mrc.py
@@ -1,6 +1,6 @@
 import mrcfile
 
-from ....core import ImageBlock
+from ....datablocks import ImageBlock
 from ...utils import guess_name
